+++ lib/ClientException.php
@@ -2,8 +2,7 @@
 declare(strict_types=1);
 namespace yii\freemobile;
 
-use GuzzleHttp\Psr7\{Uri};
-use Psr\Http\Message\{UriInterface};
+use League\Uri\{Http as Uri};
 
 /**
  * An exception caused by an error in a `Client` request.
@@ -11,19 +10,19 @@ use Psr\Http\Message\{UriInterface};
 class ClientException extends \RuntimeException {
 
   /**
-   * @var UriInterface|null The URL of the HTTP request or response that failed.
+   * @var Uri|null The URL of the HTTP request or response that failed.
    */
   private $uri;
 
   /**
    * Creates a new client exception.
    * @param string $message A message describing the error.
-   * @param string|UriInterface $uri The URL of the HTTP request or response that failed.
+   * @param string|Uri $uri The URL of the HTTP request or response that failed.
    * @param \Throwable $previous The previous exception used for the exception chaining.
    */
   function __construct($message, $uri = null, \Throwable $previous = null) {
     parent::__construct($message, 0, $previous);
-    $this->uri = is_string($uri) ? new Uri($uri) : $uri;
+    $this->uri = is_string($uri) ? Uri::createFromString($uri) : $uri;
   }
 
   /**
@@ -38,9 +37,9 @@ class ClientException extends \RuntimeException {
 
   /**
    * Gets the URL of the HTTP request or response that failed.
-   * @return UriInterface|null The URL of the HTTP request or response that failed.
+   * @return Uri|null The URL of the HTTP request or response that failed.
    */
-  function getUri(): ?UriInterface {
+  function getUri(): ?Uri {
     return $this->uri;
   }
 }